+++ Learn_Words_App_0_0_1/main.cpp
@@ -10,7 +10,8 @@ int main(int argc, char *argv[])
     QApplication a(argc, argv);
     auto window = std::make_unique<MainWindow>();
     window->setWindowTitle("Learn Words App");
-    window->setFixedSize(Sizes::MAIN_WINDOW_WIDTH, Sizes::MAIN_WINDOW_HEIGHT);
+    window->setFixedSize(static_cast<int>(Sizes::MAIN_WINDOW_WIDTH),
+                         static_cast<int>(Sizes::MAIN_WINDOW_HEIGHT));
     window->setStyleSheet(BACKGROUND_COLOR);
     window->setAutoFillBackground(true);
     window->show();
